Switch to go/types-based approach for generator.
Historically immutableGen has used a go/ast-based approach for doing the
hard work of working out whether a type is immutable or not. This is
expensive to maintain and error prone. And it means that it is very
difficult to properly support embedded immutable structs.

With Go 1.11 (and vgo) we will be getting a special version of
golang.org/x/tools/go/loader which is effectively a hybrid source/build
cache importer.

For now we use myitcv.io/hybridimporter which relies on go tip, hence
merging this into go1.11 branch.