cmd/immutableGen: properly handle non-exported setters. (#74)
We still don't fully support struct embedding because to do so we would
need to work out whether a field should be imported based on more than
whether it is simply exported. Because if the embedded struct belongs to
the same package. Don't need this for now so simply relying on Exported
which will, at worse, give us a subset of the actually embedded fields.